(reflects securityUtilities b009f9d5)

API changes:

- functions for checking only against local CRLs added
- more generic functions for calculating/checking ICV added
- some functions moved to appropriate module (to util.h)

Internal changes:

- loading of credentials (.p12 containers):
    - containers with **empty password** are **rejected** if a nonempty
      password is provided as parameter (i.e. the empty password can no longer
      be used to circumvent UTA-protection)
    - containers that don't have private keys encrypted with the same algorithm
      that is used by "storage" functions (e.g. `CREDENTIALS_save_dv`) are
      **rejected**
- default log level changed to `LOG_WARNING` (should have minimal impact as
  most logs state the level explicitly)

Other changes:

- some includes fixed
- various compiler warnings fixed

+++ src/connections/tls.c
@@ -29,7 +29,7 @@
 #include <certstatus/ocsp.h>
 #endif
 
-#include <operators.h>
+#include "secutils/operators.h"
 
 bool TLS_init(void)
 {
@@ -176,7 +176,7 @@ SSL_CTX* TLS_CTX_new(OPTIONAL SSL_CTX* ssl_ctx,
                 bak_flags = X509_VERIFY_PARAM_get_flags(vpm);
                 /* disable any cert status/revocation checking etc. */
                 X509_VERIFY_PARAM_clear_flags(vpm,
-                                              compl(X509_V_FLAG_USE_CHECK_TIME
+                                              compl((unsigned long)X509_V_FLAG_USE_CHECK_TIME
                                                 bitor X509_V_FLAG_NO_CHECK_TIME));
                 X509_VERIFY_PARAM_set_flags(vpm, X509_V_FLAG_NONFINAL_CHECK);
             }
